pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pydantic-ai-skills"
version = "0.0.1"
description = "A lightweight agent skill implementation for Pydantic AI"
authors = [
    { name = "Douglas Trajano", email = "douglas.trajano@outlook.com" },
]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai-slim>=0.0.1",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/dougtrajano/pydantic-ai-skills"
Source = "https://github.com/dougtrajano/pydantic-ai-skills"

[project.optional-dependencies]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-material-extensions>=1.3.1",
    "mkdocstrings[python]>=1.0.0",
    "mkdocs-tooltips>=0.1.0",
    "mkdocs-section-index>=0.3.10",
    "mkdocs-git-revision-date-localized-plugin>=1.5.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "types-PyYAML>=6.0.12",
]

[tool.ruff]
line-length = 120
target-version = "py310"
include = ["pydantic_ai_skills/**/*.py", "examples/**/*.py"]

[tool.ruff.lint]
extend-select = [
    "Q",
    "RUF100",
    "C90",
    "UP",
    "I",
    "D",
]
flake8-quotes = { inline-quotes = "single", multiline-quotes = "double" }
mccabe = { max-complexity = 15 }
ignore = [
    "D100", # ignore missing docstring in module
    "D102", # ignore missing docstring in public method
    "D104", # ignore missing docstring in public package
    "D105", # ignore missing docstring in magic methods
    "D107", # ignore missing docstring in __init__ methods
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["pydantic_ai_skills"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.per-file-ignores]
"examples/**/*.py" = ["D101", "D103"]
